+++ nginx/default.conf
@@ -1,5 +1,5 @@
 upstream api {
-    server api: 5000;
+    server api:5000;
 }
 
 server {
